Fixed lack of category identifiers
When I was changing CSS via a userscript, I noticed that a few blocks were not being affected.  Looking at the code I noticed that those specific blocks did not have the `"category": Blockly.Categories.sound,` property.  This simple change should fix the problem.  (There is one more block with this issue, and that is the "When backdrop switches to" block.  I will also fix that in a hotsecond)