Call updateMovable when changing shadow flag
Whether a block gets the blocklyDraggable class depends (in part) on if
it isn't a shadow block. So, when the shadow flag changes, we should
call updateMovable, so that that change is immediately reflected in the
block having the class or not.

This fixes an issue in Scratch where custom block arguments, which are
shadow blocks in XML when duplicated (and thus when initSvg'd ->
updateMovable'd), would not get the draggable class. This didn't
actually prevent the blocks from being moved, but it did mean
appropriate CSS for draggable blocks (particularly cursor: grab) was not
being applied. See LLK/scratch-blocks#1875.