Fix ctx menu closing on the same click it opens
Closure Library already handles the difficult work (keeping track
of coordinates). This commit changes handleContextMenu so that it
accesses the same code paths as normal mouse-ups (where the fix is
located in Closure) and provides a supporting override which makes
right-clicks "count" as the browser action button in this context.
It also provides the mouse event which was responsible for creating
the menu directly to the menu, which is required for the Closure
Library fix to work.